[REF]Move files needing merging into models * Move file * FIxe import in differents modules
