+++ prestashoperpconnect/__init__.py
@@ -23,12 +23,12 @@
 ###############################################################################
 
 #import backend
-import connector
+#import connector
 #import product
-import product_combination
-import sale
-import delivery
+#import product_combination
+#import sale
+#import delivery
 #import consumer
-import account
+#import account
 import models
 #import stock_tracking
