--- prestashoperpconnect/connector.py
@@ -1,51 +1,0 @@
-# -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Prestashoperpconnect : OpenERP-PrestaShop connector
-#    Copyright (C) 2013 Akretion (http://www.akretion.com/)
-#    @author: Alexis de Lattre <alexis.delattre@akretion.com>
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program. If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
-
-from openerp.osv import orm
-from openerp.addons.connector.connector import Environment
-from openerp.addons.connector.checkpoint import checkpoint
-
-
-def add_checkpoint(session, model_name, record_id, backend_id):
-    """ Add a row in the model ``connector.checkpoint`` for a record,
-    meaning it has to be reviewed by a user.
-
-    :param session: current session
-    :type session: :py:class:`openerp.addons.connector.session.ConnectorSession`
-    :param model_name: name of the model of the record to be reviewed
-    :type model_name: str
-    :param record_id: ID of the record to be reviewed
-    :type record_id: int
-    :param backend_id: ID of the Prestashop Backend
-    :type backend_id: int
-    """
-    return checkpoint.add_checkpoint(session, model_name, record_id,
-                                     'prestashop.backend', backend_id)
-
-
-def get_environment(session, model_name, backend_id):
-    model = session.pool.get('prestashop.backend')
-    backend_record = model.browse(session.cr,
-                                  session.uid,
-                                  backend_id,
-                                  session.context)
-    return Environment(backend_record, session, model_name)
