+++ prestashoperpconnect/models/consumer.py
@@ -27,7 +27,7 @@ from openerp.addons.connector.event import (on_record_write,
                                             )
 from openerp.addons.connector.connector import Binder
 from ..unit.delete_synchronizer import export_delete_record
-from ..connector import get_environment
+from .connector import get_environment
 
 
 def delay_export(session, model_name, record_id, fields=None):
