+++ prestashoperpconnect/models/delivery.py
@@ -22,6 +22,15 @@
 
 import logging
 from openerp.osv import fields, orm
+from openerp.addons.connector.unit.mapper import (mapping,
+                                                  ImportMapper,
+                                                  )
+from ..unit.backend_adapter import GenericAdapter
+from ..unit.mapper import PrestashopImportMapper
+from ..unit.import_synchronizer import (DelayedBatchImport,
+                                       PrestashopImportSynchronizer,
+                                       )
+from .backend import prestashop
 
 _logger = logging.getLogger(__name__)
 
@@ -80,3 +89,90 @@ class delivery_carrier(orm.Model):
             'res.company', 'Company', select=1, required=True
         ),
     }
+
+
+#BACKEND
+
+
+@prestashop
+class DeliveryCarrierAdapter(GenericAdapter):
+    _model_name = 'prestashop.delivery.carrier'
+    _prestashop_model = 'carriers'
+
+    def search(self, filters=None):
+        if filters is None:
+            filters = {}
+        filters['filter[deleted]'] = 0
+
+        return super(DeliveryCarrierAdapter, self).search(filters)
+
+
+@prestashop
+class DeliveryCarrierImport(PrestashopImportSynchronizer):
+    _model_name = ['prestashop.delivery.carrier']
+
+
+@prestashop
+class CarrierImportMapper(PrestashopImportMapper):
+    _model_name = 'prestashop.delivery.carrier'
+    direct = [
+        ('name', 'name_ext'),
+        ('name', 'name'),
+        ('id_reference', 'id_reference'),
+    ]
+
+    @mapping
+    def active(self, record):
+        return {'active_ext': record['active'] == '1'}
+
+    @mapping
+    def product_id(self, record):
+        if self.backend_record.shipping_product_id:
+            return {'product_id': self.backend_record.shipping_product_id.id}
+        prod_mod = self.session.pool['product.product']
+        default_ship_product = prod_mod.search(
+            self.session.cr,
+            self.session.uid,
+            [('default_code', '=', 'SHIP'),
+             ('company_id', '=', self.backend_record.company_id.id)],
+        )
+        if default_ship_product:
+            return {'product_id': default_ship_product[0]}
+        return {}
+
+    @mapping
+    def partner_id(self, record):
+        partner_pool = self.session.pool['res.partner']
+        default_partner = partner_pool.search(
+            self.session.cr,
+            self.session.uid,
+            [],
+        )[0]
+        return {'partner_id': default_partner}
+
+    @mapping
+    def prestashop_id(self, record):
+        return {'prestashop_id': record['id']}
+
+    @mapping
+    def backend_id(self, record):
+        return {'backend_id': self.backend_record.id}
+
+    @mapping
+    def company_id(self, record):
+        return {'company_id': self.backend_record.company_id.id}
+
+
+@prestashop
+class DeliveryCarrierBatchImport(DelayedBatchImport):
+    """ Import the Prestashop Carriers.
+    """
+    _model_name = ['prestashop.delivery.carrier']
+
+    def run(self, filters=None, **kwargs):
+        """ Run the synchronization """
+        record_ids = self.backend_adapter.search()
+        _logger.info('search for prestashop carriers %s returned %s',
+                     filters, record_ids)
+        for record_id in record_ids:
+            self._import_record(record_id, **kwargs)
