+++ prestashoperpconnect_catalog_manager/models/product.py
@@ -41,7 +41,7 @@ from openerp.addons.prestashoperpconnect.unit.mapper import \
 from openerp.addons.prestashoperpconnect.unit.mapper import \
     PrestashopExportMapper
 
-from openerp.addons.prestashoperpconnect.connector import get_environment
+from openerp.addons.prestashoperpconnect.models.connector import get_environment
 from openerp.addons.prestashoperpconnect.models.backend import prestashop
 from openerp.addons.prestashoperpconnect.models.product_backend import INVENTORY_FIELDS
 from .product_combination \
